Personnel Registration System with C#
A Personnel Record Program that I have prepared with C# includes features such as adding new records, updating information with a single button, deleting records with ease, updating personnel data, displaying the distribution of personnel across cities, generating salary graphs based on professions, and providing statistics on the total number of personnel. The program also includes a statistical layout showcasing the total salary expenses, as well as averages. It is designed to offer a comprehensive solution for managing personnel data.